separated Test class
